clean up paths and layouts
