+++ web/ui/src/app/(main)/contacts/page.tsx
@@ -1,7 +1,7 @@
+import { siteConfig } from "@/app/siteConfig"
 import { Button } from "@/components/Button"
 import { ArrowAnimated } from "@/components/ui/icons/ArrowAnimated"
 import { TremorPlaceholder } from "@/components/ui/icons/TremorPlaceholder"
-import { siteConfig } from "../siteConfig"
 
 export default function Settings() {
   return (
