+++ web/ui/src/app/overview/layout.tsx
@@ -1,0 +1,13 @@
+export default function Layout({
+  children,
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  return (
+    <div className="relative">
+      <main className="p-4 sm:px-6 sm:pb-10 sm:pt-10 lg:px-10 lg:pt-7">
+        {children}
+      </main>
+    </div>
+  )
+}
