--- web/ui/src/app/settings/layout.tsx
@@ -1,14 +1,0 @@
-export default function Layout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
-  return (
-    <div className="p-4 sm:px-6 sm:pb-10 sm:pt-10 lg:px-10 lg:pt-7">
-      <h1 className="text-lg font-semibold text-gray-900 sm:text-xl dark:text-gray-50">
-        Settings
-      </h1>
-      <div>{children}</div>
-    </div>
-  )
-}
